attempted to merge slow.cpp and fast.cpp into core.cpp
